Исправлена схема создания и модель пользователя

+++ src/telegram_bot/components/user/model.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from uuid import uuid4
 
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, UUID
+from sqlalchemy import Column, String, Boolean, DateTime, UUID, BigInteger
 
 from db.connection import Base
 
@@ -10,7 +10,7 @@ class TelegramUserModel(Base):
     __tablename__ = "telegram_users"
 
     uuid = Column(UUID(as_uuid=True), primary_key=True, default=uuid4)
-    id = Column(Integer, nullable=False)
+    id = Column(BigInteger, nullable=False)
     name = Column(String, nullable=False)
     username = Column(String, nullable=True, default=None)
     is_active = Column(Boolean, default=False)
